Add a global validate function (#152)
Having used this library for a while, I think it's finally time to add
this function. It's what 99.99% of users want to do, and having to pass
around a `protovalidate.Validator` really messes with a ton of functions
downstream. There's no actual state that we care about being shared,
other than a cache, and I think this is an exception to our no-global
rule.

If there's major downsides, let's hear them, but I think this is an
extremely user-favorable move.