Implement structured field and rule paths, add field and rule values to ValidationErrors (#154)
This implements the structured field and rule paths, passing the
conformance test changes in
https://github.com/bufbuild/protovalidate/pull/265.

In addition, it also adds the ability to get the field and rule values.
These are not present in the protobuf form of the violation.

The API is still heavily centered around the protobuf form of
violations, but the `ValidationError` type is no longer an alias to the
protobuf `buf.validate.Violations` message. The intention is that users
needing access to the error information will still use the proto
violations message, only using the "native" interface methods to access
in-memory data that is not feasible to transport over the wire. (e.g.
`protoreflect.Value` pointers.)

DO NOT MERGE until the following is done:
- [x] https://github.com/bufbuild/protovalidate/pull/265 is merged
- [x] Dependencies updated to stable version release