Implement Filter interface (#185)
Adds `ValidationOption` for options passed to the `Validate` function
rather than on creation of a validator.

Adds `WithFilter` option for setting a filter to control which fields
are validated.

Adds `Filter` interface:

```go
type Filter interface {
	ShouldValidate(message protoreflect.Message, descriptor protoreflect.Descriptor) bool
}
```

- `ShouldValidate` is called once for a message, once for each of its
oneofs, and once for each of its fields. The `message` value is the
message itself, and the `descriptor` value is the descriptor of the
element being operated on.
- `ShouldValidate` will be called for every oneof and field even if it
returns false for any message, field or oneof.
- Sub-messages nested under fields will _not_ be checked if
`ShouldValidate` returns false.

This is to satisfy the following usecases
- For `buf lint`, we want to be able to single out an individual field
to check examples. This can be done by using `descriptor ==
fieldDescriptor` as a filter. This can also filter oneof and message
constraints, which is desirable.
- For field masking, it is possible for a stateful filter to perform
field masking using this interface. The only important behavioral detail
for that is that we nest depth-first into submessages. This would allow
a user to implement out-of-band field masks.