Add test for CompliationError and Filter behaviours
This adds some tests around the expected behaviours of
`CompilationError` and `Filter`:

`CompilationErrorNoViolations` tests that we expect to get a
`CompilationError` back from invalid constraints defined in the source
file, in this case, a mismatch in field type and constraint type. This
is currently behaving as expected and the test passes.

`CompilationErrorWithViolations` tests that in the case where a
`CompilationError` exists and a different field also violates the
constraint, we are only returning the `CompilationError` that was found.
This is behaving as expected and the test passes.

`FilterIncludeCompilationError` tests that if there is a `CompilationError`
on a field and we set a filter for the field with the
`CompilationError`, we should get a `CompilationError` in return. This
is currently not the behaviour, since we return early in the case of a
`CompilationError`, and then we go to evaluate the message. Since the
`CompilationError` is set on the message evaluator, and the message
descriptor is not the field descriptor we are filtering for, the
`CompilationError` is lost.

`FilterExcludeCompilationError` tests that if a filter excludes a
field with a `CompilationError`, we should still return the other
violations. This is currently not behaving as expected depending on the
field ordering due to the early return behaviour -- we may not have compiled
the constraint for the field, and so it is not evaluated.