Handle compilation error (#205)
Instead of setting `CompilationError`s found in the build process on the
mesasge evaluator, this sets them on their respective field evaluators.
They are then returned when the field is evaluated. This ensures that
all constraints ae built, even when a `CompilationError` is encountered.

This makes the `Filter` API easier to use because then all constraints
are built, and then `CompilationError`s are returned during the
evaluation layer.

The main change is that `CompilationError`s are no longer "fail fast"
immediately, and instead allow for all constraints to be built and
cached.

Fixes #204