Upgrade to Protovalidate 0.10.4 (#214)
This upgrades to the Protovalidate 0.10.4 release. In doing so, it adds
a few additional tests to the expected failures list.

Note that the four failing `is_ip` changes are due to a bug in the
`isIP` validation logic and will be fixed in a follow-up PR.

This will be merged into the `sayers/upgrade_conformance` dev branch.