Widen 32-bit types in map key bindings (#219)
When binding `this` for a map, protovalidate-go uses a `map[any]any` as
the binding value. The issue with this is if a 32-bit integer type is
used for the map key, some cel expressions will fail since cel-go uses
64-bit integers. Consider:

`this.all(k, this[k] == 1)`

This results in cel-go trying a lookup of a 32-bit integer key in a
`map[any]any` with a 64-bit integer value. Therefore the value isn't
found and this results in an error such as `no such key 42`.

The long term fix here would be to properly use a cel `pb.Map`, but that
is proving quite difficult right now due to the exposed cel-go API for
working with maps. As a short-term fix, we are widening 32-bit map key
types to explicit 64-bit types when building the map binding.