Fix Ignore Always detection (#221)
This fixes a small bug in how an `IGNORE_ALWAYS` annotation is
processed.

Previously, when `IGNORE_ALWAYS` was detected, the builder just skipped
all the `processXXX` methods for a value. However, the field's
`EvaluateMessage` was still being called and there was no check present
to see if we should skip validation. Instead, it would continue
evaluating and since there was no value present, it would fail
validation.

This adds a new `IgnoreAlways` attribute on the `field` struct so that
at evaluation time, we can determine whether everything should be
skipped entirely. In addition, it renames `shouldSkip` to
`shouldIgnoreAlways` for consistency.

Note: tested this against Protovalidate 0.10.4 and can confirm that all
conformance tests pass.