Implement getField CEL function (#225)
I think it has to be a global overload, a member overload would require
a generic on the receiver which doesn't seem possible.

I'm proposing this as an eventual replacement (before 1.0) of our hack
around the fact that the `in` identifier is reserved in CEL. This is
especially urgent for protovalidate-cc which is currently carrying
patches to the CEL implementation in order to enable it, since cel-cpp
doesn't allow this sort of functionality to be added in at runtime.

Protovalidate PR: https://github.com/bufbuild/protovalidate/pull/352