Upgrade to Protovalidate v0.11.0 (#227)
This upgrades to the latest v0.11.0 release of Protovalidate. As a
result, it conforms to the convention of using `rule` everywhere instead
of `constraint`. All usages of `constraint` have been removed.

In addition, this adds a few additional tests to the expected_failures
list as a result of v0.11.0.

Also, there are two failing unit tests as a result of the upgrade, which
will be fixed in a follow-up as they seem to be related to the failing
conformance tests. These failing unit tests are marked with a `TODO`.

---------

Co-authored-by: Nick Snyder <nsnyder@buf.build>